selenium-python/requirements.txt
pytest==8.3.5
selenium==4.21.0
allure-pytest==2.15.3
allure-python-commons==2.15.3

